rearranged test packages
