+++ data/csv/avg_cost_to_insert_into_arraylist.csv
@@ -1,300 +1,300 @@
 2,1
 3,2
-4,3
+4,1
 5,3
-6,3
-7,2
-8,2
+6,1
+7,4
+8,1
 9,5
-10,5
+10,7
 11,0
-12,1
-13,7
-14,11
-15,1
-16,10
-17,14
-18,17
-19,3
+12,11
+13,1
+14,12
+15,11
+16,9
+17,9
+18,2
+19,15
 20,0
-21,6
-22,21
-23,13
-24,8
-25,23
-26,7
-27,12
-28,20
-29,11
-30,7
+21,10
+22,8
+23,10
+24,1
+25,5
+26,24
+27,24
+28,10
+29,6
+30,10
 31,12
-32,23
-33,25
+32,16
+33,22
 34,3
-35,20
+35,14
 36,0
-37,21
-38,7
-39,38
-40,14
-41,13
-42,30
-43,4
-44,18
-45,30
-46,6
-47,17
-48,27
-49,8
-50,41
-51,22
-52,29
+37,1
+38,14
+39,11
+40,30
+41,31
+42,6
+43,7
+44,11
+45,6
+46,16
+47,36
+48,4
+49,4
+50,25
+51,43
+52,10
 53,19
-54,34
-55,42
-56,49
-57,11
-58,57
-59,8
-60,35
-61,40
-62,55
-63,26
-64,62
+54,39
+55,12
+56,30
+57,18
+58,40
+59,46
+60,25
+61,47
+62,49
+63,34
+64,61
 65,0
-66,12
-67,3
-68,61
-69,41
-70,23
-71,32
-72,3
-73,49
-74,41
-75,73
-76,28
-77,61
-78,11
-79,43
-80,35
+66,53
+67,12
+68,35
+69,17
+70,63
+71,53
+72,65
+73,35
+74,10
+75,53
+76,35
+77,73
+78,17
+79,38
+80,76
 81,57
-82,43
-83,48
-84,67
-85,64
-86,35
-87,31
-88,14
-89,35
-90,17
-91,62
-92,73
-93,49
-94,48
-95,70
-96,35
-97,49
-98,24
-99,47
-100,26
-101,22
-102,95
-103,40
-104,70
-105,58
-106,66
-107,30
-108,7
-109,92
-110,22
-111,104
-112,59
-113,95
-114,25
-115,93
-116,30
+82,45
+83,52
+84,75
+85,4
+86,42
+87,23
+88,77
+89,60
+90,49
+91,39
+92,25
+93,75
+94,8
+95,1
+96,49
+97,25
+98,80
+99,77
+100,58
+101,91
+102,49
+103,24
+104,6
+105,103
+106,34
+107,83
+108,49
+109,54
+110,41
+111,85
+112,5
+113,71
+114,109
+115,81
+116,77
 117,0
-118,81
-119,41
-120,11
-121,53
-122,84
-123,40
-124,51
-125,14
-126,12
-127,53
-128,24
-129,34
-130,59
-131,12
-132,118
-133,127
-134,123
-135,10
-136,5
-137,75
-138,48
-139,30
-140,139
-141,137
-142,133
-143,10
-144,141
+118,115
+119,32
+120,33
+121,4
+122,13
+123,2
+124,85
+125,51
+126,125
+127,16
+128,52
+129,11
+130,69
+131,7
+132,69
+133,77
+134,46
+135,86
+136,22
+137,48
+138,110
+139,95
+140,89
+141,17
+142,126
+143,84
+144,35
 145,1
-146,120
-147,119
-148,84
-149,113
-150,147
-151,134
-152,119
-153,38
-154,33
-155,137
-156,50
-157,144
-158,4
-159,27
-160,68
-161,158
-162,112
-163,59
-164,24
-165,158
-166,140
-167,155
-168,40
-169,110
-170,157
-171,117
-172,12
-173,82
-174,44
-175,35
-176,80
-177,14
-178,67
-179,163
-180,161
-181,9
-182,164
-183,146
-184,174
-185,95
-186,26
-187,141
-188,5
-189,111
-190,144
-191,100
-192,138
-193,175
-194,122
-195,115
-196,66
-197,150
-198,147
-199,156
-200,175
-201,66
-202,198
-203,151
-204,174
-205,38
-206,169
-207,139
-208,181
-209,138
+146,47
+147,134
+148,69
+149,93
+150,120
+151,26
+152,95
+153,131
+154,52
+155,135
+156,148
+157,102
+158,129
+159,137
+160,23
+161,50
+162,102
+163,44
+164,163
+165,146
+166,110
+167,37
+168,64
+169,29
+170,118
+171,120
+172,33
+173,128
+174,51
+175,119
+176,167
+177,47
+178,77
+179,59
+180,131
+181,130
+182,33
+183,151
+184,124
+185,42
+186,44
+187,66
+188,95
+189,188
+190,26
+191,173
+192,116
+193,58
+194,157
+195,114
+196,9
+197,66
+198,6
+199,15
+200,115
+201,148
+202,150
+203,8
+204,118
+205,47
+206,36
+207,10
+208,62
+209,179
 210,0
-211,105
-212,50
-213,209
-214,84
-215,47
-216,109
-217,6
-218,121
-219,11
-220,13
-221,52
-222,46
-223,217
-224,152
-225,148
-226,154
-227,158
-228,182
-229,104
-230,105
-231,193
-232,164
-233,194
-234,112
-235,197
-236,128
-237,48
-238,2
-239,208
-240,34
-241,101
-242,107
-243,146
-244,116
-245,134
-246,140
-247,14
-248,144
-249,210
-250,157
-251,115
-252,87
-253,28
-254,121
-255,240
-256,116
-257,34
-258,136
-259,78
-260,87
-261,70
-262,175
-263,94
-264,102
-265,217
-266,47
-267,163
-268,164
-269,93
-270,84
-271,12
-272,228
-273,51
-274,15
-275,30
-276,191
-277,242
-278,248
-279,131
-280,270
-281,263
-282,16
-283,233
-284,220
-285,78
-286,245
-287,41
-288,155
-289,17
-290,175
-291,226
-292,77
-293,91
-294,194
-295,89
-296,151
-297,99
-298,52
-299,77
-300,29
-301,101
+211,58
+212,41
+213,212
+214,159
+215,8
+216,5
+217,180
+218,139
+219,145
+220,212
+221,144
+222,163
+223,58
+224,141
+225,40
+226,4
+227,53
+228,91
+229,205
+230,72
+231,34
+232,62
+233,132
+234,162
+235,161
+236,118
+237,142
+238,20
+239,106
+240,116
+241,6
+242,208
+243,187
+244,13
+245,133
+246,105
+247,119
+248,76
+249,124
+250,171
+251,67
+252,173
+253,186
+254,40
+255,127
+256,8
+257,248
+258,210
+259,201
+260,227
+261,115
+262,185
+263,120
+264,258
+265,58
+266,228
+267,159
+268,22
+269,203
+270,54
+271,192
+272,250
+273,188
+274,141
+275,16
+276,142
+277,139
+278,237
+279,164
+280,213
+281,25
+282,256
+283,197
+284,122
+285,174
+286,113
+287,121
+288,131
+289,88
+290,129
+291,49
+292,42
+293,73
+294,64
+295,254
+296,158
+297,98
+298,110
+299,160
+300,112
+301,43
