+++ data/csv/avg_cost_to_remove_from_arraylist.csv
@@ -1,300 +1,300 @@
-300,189
-299,226
-298,193
-297,285
-296,181
-295,215
-294,200
-293,37
-292,237
-291,98
-290,65
-289,189
-288,135
-287,96
-286,222
-285,161
-284,84
-283,172
-282,222
-281,227
-280,221
-279,191
-278,233
-277,131
-276,270
-275,42
-274,128
-273,175
-272,107
-271,130
-270,258
-269,268
-268,51
-267,240
-266,158
-265,53
-264,56
-263,40
-262,191
-261,31
-260,169
-259,60
-258,27
-257,143
-256,136
-255,255
-254,215
-253,54
-252,216
-251,75
-250,153
-249,233
-248,226
-247,102
-246,103
-245,175
-244,228
-243,163
-242,76
-241,10
-240,163
-239,31
-238,189
-237,214
-236,169
-235,58
-234,223
-233,189
-232,92
-231,224
-230,114
-229,91
-228,156
-227,163
-226,183
-225,7
-224,66
-223,59
-222,207
-221,79
-220,53
-219,211
-218,207
-217,72
-216,201
-215,200
-214,55
-213,101
-212,133
-211,187
-210,128
-209,36
-208,87
-207,42
-206,5
-205,64
-204,196
-203,195
-202,61
-201,162
-200,17
-199,54
-198,85
-197,30
-196,40
-195,186
-194,83
-193,33
-192,116
-191,77
-190,150
-189,81
-188,19
-187,147
-186,59
-185,79
-184,94
-183,90
-182,24
-181,97
-180,29
-179,62
-178,96
-177,72
-176,59
-175,55
-174,114
-173,78
-172,4
-171,118
-170,9
-169,149
-168,113
-167,13
-166,29
-165,39
-164,107
-163,30
-162,109
-161,135
-160,62
-159,79
-158,93
-157,87
-156,128
-155,2
-154,93
-153,72
-152,98
-151,75
-150,132
-149,104
-148,128
-147,36
-146,73
-145,118
-144,52
-143,4
-142,71
-141,49
-140,18
-139,38
-138,35
-137,39
-136,134
-135,37
-134,102
-133,33
-132,59
-131,78
-130,31
-129,56
-128,127
-127,27
-126,102
-125,31
-124,61
-123,91
-122,111
-121,59
-120,111
-119,6
-118,89
-117,77
-116,31
-115,11
-114,16
-113,86
-112,96
-111,108
-110,106
-109,103
-108,2
-107,70
-106,104
-105,4
-104,75
-103,38
-102,52
+300,175
+299,21
+298,123
+297,226
+296,52
+295,190
+294,272
+293,127
+292,182
+291,97
+290,218
+289,16
+288,19
+287,32
+286,132
+285,273
+284,87
+283,227
+282,92
+281,173
+280,17
+279,84
+278,218
+277,184
+276,180
+275,192
+274,215
+273,68
+272,217
+271,95
+270,211
+269,10
+268,226
+267,145
+266,105
+265,205
+264,195
+263,36
+262,221
+261,163
+260,199
+259,99
+258,90
+257,94
+256,176
+255,230
+254,103
+253,180
+252,53
+251,208
+250,10
+249,135
+248,40
+247,174
+246,132
+245,162
+244,113
+243,146
+242,171
+241,34
+240,125
+239,188
+238,8
+237,136
+236,135
+235,166
+234,207
+233,81
+232,225
+231,167
+230,217
+229,218
+228,175
+227,223
+226,134
+225,214
+224,171
+223,172
+222,31
+221,32
+220,128
+219,152
+218,135
+217,107
+216,45
+215,64
+214,170
+213,28
+212,5
+211,94
+210,14
+209,121
+208,179
+207,102
+206,179
+205,188
+204,71
+203,134
+202,86
+201,136
+200,172
+199,117
+198,195
+197,68
+196,166
+195,15
+194,166
+193,73
+192,94
+191,108
+190,23
+189,76
+188,136
+187,83
+186,50
+185,55
+184,50
+183,118
+182,64
+181,104
+180,11
+179,42
+178,35
+177,68
+176,117
+175,94
+174,157
+173,53
+172,155
+171,21
+170,102
+169,40
+168,62
+167,60
+166,44
+165,152
+164,115
+163,78
+162,56
+161,94
+160,20
+159,30
+158,24
+157,111
+156,92
+155,44
+154,77
+153,63
+152,76
+151,38
+150,24
+149,122
+148,7
+147,73
+146,75
+145,78
+144,21
+143,102
+142,40
+141,53
+140,53
+139,123
+138,21
+137,25
+136,108
+135,18
+134,128
+133,20
+132,91
+131,69
+130,59
+129,119
+128,102
+127,59
+126,24
+125,120
+124,24
+123,88
+122,90
+121,81
+120,93
+119,3
+118,27
+117,108
+116,7
+115,91
+114,114
+113,21
+112,3
+111,12
+110,16
+109,82
+108,46
+107,98
+106,17
+105,83
+104,87
+103,51
+102,88
 101,96
-100,70
-99,5
-98,66
-97,12
-96,17
-95,90
-94,6
-93,35
-92,47
-91,6
-90,32
-89,45
-88,50
-87,45
-86,38
-85,26
-84,11
-83,68
-82,25
-81,32
-80,16
-79,29
-78,39
-77,37
-76,7
-75,15
-74,73
-73,5
-72,6
-71,18
-70,67
-69,3
-68,28
-67,56
-66,5
-65,21
-64,39
+100,75
+99,73
+98,57
+97,97
+96,60
+95,87
+94,15
+93,15
+92,72
+91,33
+90,31
+89,22
+88,43
+87,14
+86,29
+85,35
+84,55
+83,71
+82,28
+81,15
+80,38
+79,66
+78,45
+77,63
+76,69
+75,55
+74,71
+73,38
+72,62
+71,35
+70,59
+69,53
+68,52
+67,8
+66,19
+65,40
+64,59
 63,11
-62,44
-61,59
-60,45
-59,49
-58,52
-57,35
-56,3
-55,50
-54,39
-53,21
-52,21
-51,18
-50,1
-49,6
-48,44
-47,17
-46,6
-45,45
-44,4
-43,9
-42,9
-41,41
-40,36
-39,37
-38,2
-37,8
-36,19
+62,29
+61,4
+60,54
+59,23
+58,17
+57,42
+56,19
+55,8
+54,9
+53,51
+52,31
+51,36
+50,34
+49,26
+48,17
+47,24
+46,45
+45,34
+44,10
+43,43
+42,41
+41,26
+40,13
+39,4
+38,7
+37,1
+36,7
 35,3
-34,29
-33,30
-32,13
-31,30
-30,22
-29,5
-28,20
+34,9
+33,31
+32,21
+31,2
+30,17
+29,27
+28,28
 27,22
-26,14
-25,7
-24,20
-23,6
-22,21
-21,1
-20,1
-19,14
-18,12
-17,2
-16,11
-15,11
-14,6
-13,5
-12,3
-11,2
-10,3
-9,5
-8,6
+26,26
+25,5
+24,21
+23,16
+22,19
+21,10
+20,8
+19,8
+18,1
+17,9
+16,9
+15,2
+14,5
+13,7
+12,7
+11,3
+10,2
+9,9
+8,5
 7,5
 6,3
-5,1
-4,1
-3,3
+5,2
+4,2
+3,1
 2,1
 1,1
