+++ data/csv/levels_in_avl_tree_vs_n_random_keys.csv
@@ -4,7 +4,7 @@
 4,3
 5,3
 6,3
-7,4
+7,3
 8,4
 9,4
 10,4
@@ -19,7 +19,6 @@
 19,5
 20,5
 21,5
-22,5
 22,5
 23,6
 24,6
@@ -38,14 +37,14 @@
 37,6
 38,6
 39,6
-40,6
-41,6
-42,6
-43,6
-44,6
-45,6
-46,6
-47,6
+40,7
+41,7
+42,7
+43,7
+44,7
+45,7
+46,7
+47,7
 48,7
 49,7
 50,7
@@ -78,18 +77,17 @@
 77,7
 78,7
 79,7
-80,7
-81,7
-82,7
-83,7
-84,7
-85,7
-86,7
-87,8
-88,8
+80,8
+81,8
+82,8
+83,8
+84,8
+85,8
+86,8
+87,7
+88,7
 89,8
 90,8
-91,8
 91,8
 92,8
 93,8
@@ -98,3 +96,5 @@
 96,8
 97,8
 98,8
+99,8
+100,8
