+++ data/csv/levels_in_bst_tree_vs_n_random_keys.csv
@@ -3,27 +3,27 @@
 3,3
 4,4
 5,4
-6,5
-7,6
-8,6
-9,7
-10,7
-11,7
-12,7
+6,4
+7,5
+8,5
+9,5
+10,5
+11,6
+12,6
 13,7
-14,7
-15,7
-16,7
-17,7
-18,7
+14,8
+15,8
+16,8
+17,8
+18,8
 19,8
 20,8
-21,8
+21,9
 22,9
 23,9
 24,9
 25,9
-26,10
+26,9
 27,10
 28,10
 29,10
@@ -31,70 +31,70 @@
 31,10
 32,10
 33,10
-34,10
-35,10
-36,10
-37,10
-38,10
-39,10
-40,10
-41,10
-42,10
-43,10
-44,10
-45,10
-46,10
-47,10
-48,10
-49,10
-50,10
-51,10
-52,11
-53,11
-54,11
+34,11
+35,11
+36,11
+37,11
+38,11
+39,11
+40,12
+41,12
+42,12
+43,12
+44,12
+45,12
+46,12
+47,12
+48,12
+49,12
+50,12
+51,12
+52,12
+53,12
+54,12
 55,12
 56,12
-57,12
-58,12
-59,12
-60,12
-61,12
-62,12
-63,12
-64,12
-65,12
-66,12
-67,12
-68,12
-69,12
-70,12
-71,12
-72,12
-73,12
-74,12
-75,12
-76,12
-77,12
-78,12
-79,12
-80,12
-81,12
-82,12
-83,13
-84,13
-85,13
-86,13
-87,14
-88,14
-89,14
-90,14
-91,14
-92,14
-93,14
-94,14
-95,14
-96,14
-97,14
-98,14
-99,14
-100,14
+57,13
+58,13
+58,13
+59,13
+60,13
+61,13
+62,13
+63,14
+64,14
+65,14
+66,14
+67,14
+68,14
+69,14
+70,14
+71,14
+72,14
+73,14
+74,14
+74,14
+75,14
+76,14
+77,14
+78,14
+79,14
+80,14
+81,14
+82,14
+83,14
+84,14
+85,14
+86,15
+87,15
+88,15
+89,15
+90,15
+91,15
+92,15
+93,15
+94,15
+95,15
+96,15
+97,15
+98,15
