+++ data/csv/shifts_required_when_inserting_rand_index.csv
@@ -1,300 +1,300 @@
 2,1
 3,3
 4,4
-5,5
-6,10
-7,12
-8,16
-9,22
+5,8
+6,11
+7,16
+8,20
+9,23
 10,26
 11,26
-12,31
-13,35
-14,36
-15,42
-16,49
-17,51
-18,64
-19,72
-20,72
-21,79
-22,95
-23,114
-24,122
-25,137
-26,147
-27,171
-28,185
-29,207
-30,234
-31,251
-32,262
-33,286
-34,287
-35,309
-36,309
-37,318
-38,326
-39,359
-40,375
-41,412
-42,436
-43,453
-44,475
-45,512
-46,514
-47,546
-48,586
-49,616
-50,619
-51,633
-52,661
-53,703
-54,742
-55,767
-56,787
-57,843
-58,863
+12,33
+13,42
+14,49
+15,50
+16,54
+17,68
+18,77
+19,93
+20,93
+21,113
+22,134
+23,135
+24,136
+25,145
+26,165
+27,186
+28,211
+29,231
+30,259
+31,271
+32,300
+33,316
+34,345
+35,377
+36,377
+37,409
+38,443
+39,450
+40,459
+41,462
+42,495
+43,504
+44,518
+45,561
+46,570
+47,604
+48,648
+49,656
+50,693
+51,707
+52,741
+53,747
+54,765
+55,811
+56,838
+57,883
+58,896
 59,907
-60,914
-61,936
-62,971
-63,1003
-64,1006
-65,1006
-66,1052
-67,1073
-68,1133
-69,1166
-70,1199
-71,1202
-72,1243
-73,1262
-74,1289
-75,1333
-76,1387
-77,1388
-78,1405
-79,1456
-80,1467
-81,1509
-82,1574
-83,1609
-84,1631
-85,1654
-86,1714
-87,1720
-88,1727
-89,1743
-90,1757
-91,1829
-92,1912
-93,1940
-94,1998
-95,2041
-96,2073
-97,2128
-98,2173
-99,2203
-100,2252
-101,2351
-102,2369
-103,2448
-104,2460
-105,2524
-106,2599
-107,2633
-108,2655
-109,2693
-110,2759
-111,2802
-112,2829
-113,2897
-114,2908
-115,2991
-116,3059
-117,3059
-118,3168
-119,3216
-120,3230
-121,3266
-122,3291
-123,3329
-124,3372
-125,3427
-126,3534
-127,3617
-128,3618
-129,3678
-130,3773
-131,3856
-132,3933
-133,3953
-134,4025
-135,4149
-136,4192
-137,4281
-138,4336
-139,4422
-140,4519
-141,4657
-142,4764
-143,4766
-144,4775
-145,4912
-146,5044
-147,5056
-148,5199
-149,5237
-150,5312
-151,5411
-152,5502
-153,5624
-154,5679
-155,5704
-156,5835
-157,5895
-158,5997
-159,6062
-160,6187
-161,6291
-162,6387
-163,6538
-164,6564
-165,6683
-166,6843
-167,6993
-168,7001
-169,7091
-170,7145
-171,7184
-172,7238
-173,7326
-174,7381
-175,7494
-176,7537
-177,7613
-178,7614
-179,7616
-180,7617
-181,7774
-182,7850
-183,7917
-184,8071
-185,8144
-186,8182
-187,8288
-188,8347
-189,8529
-190,8594
-191,8598
-192,8620
-193,8673
-194,8793
-195,8928
-196,8960
-197,9002
-198,9072
-199,9111
-200,9149
-201,9168
-202,9291
-203,9400
-204,9465
-205,9471
-206,9527
-207,9542
-208,9742
-209,9817
-210,9817
-211,10015
-212,10092
-213,10095
-214,10174
-215,10292
-216,10450
-217,10537
-218,10565
-219,10731
-220,10942
-221,11159
-222,11366
-223,11375
-224,11456
-225,11525
-226,11614
-227,11735
-228,11895
-229,11995
-230,12016
-231,12158
-232,12212
-233,12276
-234,12453
-235,12470
-236,12678
-237,12826
-238,12954
-239,13041
-240,13080
-241,13230
-242,13452
-243,13457
-244,13617
-245,13787
-246,13946
-247,13974
-248,14036
-249,14238
-250,14360
-251,14519
-252,14538
-253,14726
-254,14863
-255,15065
-256,15113
-257,15134
-258,15163
-259,15320
-260,15417
-261,15622
-262,15864
-263,15931
-264,16189
-265,16395
-266,16587
-267,16650
-268,16800
-269,17007
-270,17270
-271,17527
-272,17710
-273,17789
-274,17950
-275,18025
-276,18028
-277,18166
-278,18374
-279,18401
-280,18491
-281,18610
-282,18861
-283,18862
-284,18875
-285,19030
-286,19102
-287,19248
-288,19403
-289,19671
-290,19826
-291,19990
-292,20195
-293,20398
-294,20595
-295,20885
-296,21074
-297,21126
-298,21377
-299,21502
-300,21531
-301,21577
+60,911
+61,965
+62,980
+63,1009
+64,1065
+65,1065
+66,1129
+67,1137
+68,1180
+69,1220
+70,1231
+71,1256
+72,1284
+73,1354
+74,1384
+75,1419
+76,1442
+77,1450
+78,1513
+79,1588
+80,1622
+81,1624
+82,1648
+83,1672
+84,1734
+85,1737
+86,1797
+87,1845
+88,1916
+89,1926
+90,1966
+91,2041
+92,2056
+93,2080
+94,2087
+95,2150
+96,2201
+97,2270
+98,2302
+99,2376
+100,2440
+101,2471
+102,2511
+103,2584
+104,2620
+105,2712
+106,2723
+107,2735
+108,2737
+109,2760
+110,2801
+111,2864
+112,2942
+113,2984
+114,3082
+115,3129
+116,3241
+117,3241
+118,3352
+119,3405
+120,3442
+121,3506
+122,3549
+123,3644
+124,3713
+125,3784
+126,3791
+127,3904
+128,3914
+129,3976
+130,3990
+131,3995
+132,4126
+133,4191
+134,4257
+135,4321
+136,4389
+137,4518
+138,4616
+139,4742
+140,4814
+141,4896
+142,4944
+143,4988
+144,5052
+145,5196
+146,5268
+147,5355
+148,5445
+149,5478
+150,5605
+151,5749
+152,5818
+153,5874
+154,5918
+155,5966
+156,6108
+157,6192
+158,6242
+159,6361
+160,6500
+161,6591
+162,6658
+163,6791
+164,6948
+165,6991
+166,7024
+167,7103
+168,7116
+169,7219
+170,7334
+171,7475
+172,7619
+173,7724
+174,7876
+175,7926
+176,7933
+177,7982
+178,8102
+179,8118
+180,8147
+181,8257
+182,8266
+183,8357
+184,8500
+185,8557
+186,8710
+187,8847
+188,8874
+189,9042
+190,9191
+191,9205
+192,9309
+193,9400
+194,9560
+195,9744
+196,9881
+197,10030
+198,10188
+199,10334
+200,10456
+201,10571
+202,10765
+203,10808
+204,10941
+205,10972
+206,11040
+207,11153
+208,11170
+209,11185
+210,11185
+211,11327
+212,11529
+213,11631
+214,11752
+215,11911
+216,12121
+217,12315
+218,12506
+219,12606
+220,12700
+221,12803
+222,12926
+223,13047
+224,13164
+225,13343
+226,13555
+227,13735
+228,13916
+229,13952
+230,13981
+231,14101
+232,14150
+233,14218
+234,14391
+235,14416
+236,14642
+237,14655
+238,14716
+239,14944
+240,14969
+241,15143
+242,15193
+243,15201
+244,15302
+245,15480
+246,15524
+247,15542
+248,15676
+249,15765
+250,15929
+251,15964
+252,16108
+253,16129
+254,16165
+255,16404
+256,16620
+257,16678
+258,16775
+259,16870
+260,16988
+261,17032
+262,17289
+263,17456
+264,17691
+265,17847
+266,18065
+267,18102
+268,18202
+269,18281
+270,18333
+271,18520
+272,18728
+273,18876
+274,18879
+275,19109
+276,19190
+277,19447
+278,19666
+279,19812
+280,19836
+281,20030
+282,20103
+283,20239
+284,20258
+285,20298
+286,20379
+287,20463
+288,20632
+289,20666
+290,20830
+291,21094
+292,21164
+293,21354
+294,21601
+295,21610
+296,21698
+297,21905
+298,22098
+299,22139
+300,22146
+301,22412
