--- src/test/groovy/analytics/graphs/Graphs.groovy
@@ -1,6 +1,0 @@
-package analytics.graphs
-
-import spock.lang.Specification
-
-class Graphs extends Specification {
-}
