--- src/test/groovy/analytics/maps/Maps.groovy
@@ -1,6 +1,0 @@
-package analytics.maps
-
-import spock.lang.Specification
-
-class Maps extends Specification {
-}
