--- src/test/groovy/analytics/sets/Sets.groovy
@@ -1,6 +1,0 @@
-package analytics.sets
-
-import spock.lang.Specification
-
-class Sets extends Specification {
-}
