--- src/test/groovy/functional/graphs/Graphs.groovy
@@ -1,6 +1,0 @@
-package functional.graphs
-
-import spock.lang.Specification
-
-class Graphs extends Specification {
-}
