--- src/test/groovy/functional/maps/Maps.groovy
@@ -1,6 +1,0 @@
-package functional.maps
-
-import spock.lang.Specification
-
-class Maps extends Specification {
-}
