--- src/test/groovy/functional/sets/Sets.groovy
@@ -1,6 +1,0 @@
-package functional.sets
-
-import spock.lang.Specification
-
-class Sets extends Specification {
-}
