--- src/test/groovy/functional/trees/Trees.groovy
@@ -1,6 +1,0 @@
-package functional.trees
-
-import spock.lang.Specification
-
-class Trees extends Specification {
-}
