+++ src/test/groovy/structures/functional/graphs/Graphs.groovy
@@ -1,0 +1,6 @@
+package structures.functional.graphs
+
+import spock.lang.Specification
+
+class Graphs extends Specification {
+}
