+++ src/test/groovy/structures/functional/maps/Maps.groovy
@@ -1,0 +1,6 @@
+package structures.functional.maps
+
+import spock.lang.Specification
+
+class Maps extends Specification {
+}
