+++ src/test/groovy/structures/functional/sets/Sets.groovy
@@ -1,0 +1,6 @@
+package structures.functional.sets
+
+import spock.lang.Specification
+
+class Sets extends Specification {
+}
