+++ src/test/groovy/structures/functional/trees/Trees.groovy
@@ -1,0 +1,6 @@
+package structures.functional.trees
+
+import spock.lang.Specification
+
+class Trees extends Specification {
+}
