+++ src/test/groovy/structures/performance/graphs/Graphs.groovy
@@ -1,0 +1,6 @@
+package structures.performance.graphs
+
+import spock.lang.Specification
+
+class Graphs extends Specification {
+}
