+++ src/test/groovy/structures/performance/maps/Maps.groovy
@@ -1,0 +1,6 @@
+package structures.performance.maps
+
+import spock.lang.Specification
+
+class Maps extends Specification {
+}
