+++ src/test/groovy/structures/performance/sets/Sets.groovy
@@ -1,0 +1,6 @@
+package structures.performance.sets
+
+import spock.lang.Specification
+
+class Sets extends Specification {
+}
