+++ src/test/groovy/structures/unit/commons/DataStructure.groovy
@@ -1,0 +1,6 @@
+package structures.unit.commons
+
+import spock.lang.Specification
+
+class DataStructure extends Specification {
+}
